meu-agente-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-ai==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
openai==1.3.8
asyncio==3.4.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
griffe
